+++ crates/cxx-qt-build/src/dir.rs
@@ -7,7 +7,8 @@
 
 use qt_build_utils::QmlUri;
 
-use crate::{crate_name, module_name_from_uri};
+use crate::crate_name;
+use std::ffi::OsStr;
 use std::io::Result;
 use std::{
     env, fs,
@@ -58,19 +59,52 @@ pub(crate) fn crate_target() -> PathBuf {
     }
 }
 
-/// The target directory, namespaced by QML module
-pub(crate) fn module_target(module_uri: &QmlUri) -> PathBuf {
-    module_export(module_uri).unwrap_or_else(|| {
-        out()
-            // Use a short name due to the Windows file path limit!
-            .join("cxxqtqml")
-            .join(module_name_from_uri(module_uri))
-    })
+/// The cargo /target directory that is being used
+// Loosely based on
+// https://github.com/dtolnay/cxx/blob/5dca8106b2b218b7e85090cf7c641651ddd81537/gen/build/src/target.rs
+pub(crate) fn cargo_target_dir() -> core::result::Result<PathBuf, env::VarError> {
+    env::var("CARGO_TARGET_DIR")
+        .map(PathBuf::from)
+        .map(|path| path.join("cxxqt"))
+        .or_else(|_| {
+            // fs::canonicalize on Windows produces UNC paths which cl.exe is unable to
+            // handle in includes.
+            // https://github.com/rust-lang/rust/issues/42869
+            // https://github.com/alexcrichton/cc-rs/issues/169
+            let mut also_try_canonical = cfg!(not(windows));
+
+            let out_dir = out();
+            let mut dir = out_dir.clone();
+            loop {
+                if dir.join(".rustc_info.json").exists()
+                    || dir.join("CACHEDIR.TAG").exists()
+                    || dir.file_name() == Some(OsStr::new("target"))
+                        && dir
+                            .parent()
+                            .is_some_and(|parent| parent.join("Cargo.toml").exists())
+                {
+                    return Ok(dir.join("cxxqt"));
+                }
+                if dir.pop() {
+                    continue;
+                }
+                if also_try_canonical {
+                    if let Ok(canonical_dir) = out_dir.canonicalize() {
+                        dir = canonical_dir;
+                        also_try_canonical = false;
+                        continue;
+                    }
+                }
+
+                // We cannot find by scanning the directories so use a VarError for now
+                return Err(env::VarError::NotPresent);
+            }
+        })
 }
 
 /// The export directory, namespaced by QML module
 ///
-/// In conctrast to the crate_export directory, this is `Some` for downstream dependencies as well.
+/// In contrast to the crate_export directory, this is `Some` for downstream dependencies as well.
 /// This allows CMake to import QML modules from dependencies.
 ///
 /// TODO: This may conflict if two dependencies are building QML modules with the same name!
@@ -78,13 +112,18 @@ pub(crate) fn module_target(module_uri: &QmlUri) -> PathBuf {
 pub(crate) fn module_export(module_uri: &QmlUri) -> Option<PathBuf> {
     // In contrast to crate_export, we don't need to check for the specific crate here.
     // QML modules should always be exported.
+    module_export_qml_modules().map(|dir| dir.join(module_uri.as_dirs()))
+}
+
+pub(crate) fn module_export_qml_modules() -> Option<PathBuf> {
+    // In contrast to crate_export, we don't need to check for the specific crate here.
+    // QML modules should always be exported.
     env::var("CXX_QT_EXPORT_DIR")
-        .ok()
         .map(PathBuf::from)
-        .map(|dir| {
-            dir.join("qml_modules")
-                .join(module_name_from_uri(module_uri))
-        })
+        // Fallback to find the target/ directory when using Cargo
+        .or_else(|_| cargo_target_dir())
+        .ok()
+        .map(|dir| dir.join("qml_modules"))
 }
 
 /// The target directory or another directory where we can write files that will be shared