+++ crates/cxx-qt-lib-extras/build.rs
@@ -63,14 +63,11 @@ fn main() {
         "gui/qapplication",
     ];
 
-    builder = builder.cc_builder(move |cc| {
-        for cpp_file in &cpp_files {
-            cc.file(format!("src/{cpp_file}.cpp"));
-            println!("cargo::rerun-if-changed=src/{cpp_file}.cpp");
-        }
-        cc.file("src/qt_types.cpp");
-        println!("cargo::rerun-if-changed=src/qt_types.cpp");
-    });
+    for cpp_file in &cpp_files {
+        builder = builder.cpp_file(format!("src/{cpp_file}.cpp"));
+    }
+    builder = builder.cpp_file("src/qt_types.cpp");
+
     println!("cargo::rerun-if-changed=src/assertion_utils.h");
 
     let interface = builder