+++ examples/qml_minimal_plugin/cpp/main.cpp
@@ -1,0 +1,1 @@
+../../qml_minimal/cpp/main.cpp