+++ tests/qt_types_standalone/rust/build.rs
@@ -16,6 +16,7 @@ fn main() {
         .file("src/qdate.rs")
         .file("src/qdatetime.rs")
         .file("src/qflags.rs")
+        .file("src/qgenericmatrix.rs")
         .file("src/qguiapplication.rs")
         .file("src/qhash.rs")
         .file("src/qline.rs")